todo/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
gunicorn
whitenoise
psycopg2-binary
dj-database-url
